requirements.txt
Flask==2.2.5
flask-socketio==5.3.6
flask-cors==6.0.1
python-telegram-bot[webhooks]==20.6
python-dotenv==1.0.1
nest_asyncio==1.6.0
psycopg2-binary==2.9.9
cryptography==41.0.7
python-dateutil==2.8.2